Support for Windows acceleration curve.
Combo-boxes in prefpane is now enabled, and has two curves: Linear and Windows.

Note, last two velocity slider positions maps to the same slider position for the windows curve.
This is becuase the windows curve only has 11 slider position, we have 12.

Bumped version to 1.0.2.

Uses a very very very slimmed down version of libpointing for the "Windows transfer function".

Note that libpointing license is LGPL. This has implications on how it can be distributed, etc.